refactor(canvas-editor): extract presentation tokens to leaf module
Move PRES_COLORS/PRES_CONFIG/getPresColors/PRES_BACKGROUND_COLORS out of
configs/presentation/presentationTheme into a dependency-free leaf module
utils/presentationTokens. PresentationDesignSection (sidebar) imported
PRES_COLORS from configs — the single runtime back-edge that closed a
sidebar<->configs module-init cycle and crashed canvas-editor under natural
code-splitting (Cannot read properties of undefined reading 'map'). The old
path re-exports from the leaf, so the ~12 config-side consumers are unchanged.
